[#188] refactor build to remove unused register-common
register-common was included as a soft dependency, but this is not used
anywhere. This is in contrast to register-sources-bods, which uses it as
a soft dependency (not including in the gemspec, but requiring it to be
used by the app and also in testing), andn in contrast to
register-sources-oc, which doesn't include it as any kind of dependency.
Including it when it is not needed needlessly complicates the build.